dispatcher: refactoring dispatcher to handle new sdk
