+++ piano-dispatcher/src/main/java/com/batch/android/dispatcher/piano/PianoDispatcher.java
@@ -1,6 +1,5 @@
 package com.batch.android.dispatcher.piano;
 
-import android.content.Context;
 import android.net.Uri;
 
 import androidx.annotation.NonNull;
@@ -14,45 +13,43 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
-import io.piano.analytics.Event;
-import io.piano.analytics.PianoAnalytics;
 
 /**
  * Piano Event Dispatcher
- *
+ * <p>
  * Dispatch Batch events to the Piano Analytics SDK. By default events are dispatched as On-site Ads.
- * If you want to dispatch as custom event, please see {@link PianoDispatcher#enableBatchCustomEvents(boolean)}.
+ * If you want to dispatch as custom event, please see {@link LegacyPianoDispatcher#enableBatchCustomEvents(boolean)}.
  * Note: if you enable custom events, you need to declare them in your Piano Data Model.
  */
-public class PianoDispatcher implements BatchEventDispatcher {
+public abstract class PianoDispatcher implements BatchEventDispatcher {
 
     /**
      * Batch internal dispatcher information used for analytics
      */
     private static final String DISPATCHER_NAME = "piano";
-    private static final int DISPATCHER_VERSION = 1;
+    private static final int DISPATCHER_VERSION = 2;
 
     /**
      * Piano event keys
      */
-    private static final String CAMPAIGN = "src_campaign";
-    private static final String SOURCE = "src_source";
-    private static final String SOURCE_FORCE = "src_force";
-    private static final String MEDIUM = "src_medium";
-    private static final String CONTENT = "src_content";
+    protected static final String CAMPAIGN = "src_campaign";
+    protected static final String SOURCE = "src_source";
+    protected static final String SOURCE_FORCE = "src_force";
+    protected static final String MEDIUM = "src_medium";
+    protected static final String CONTENT = "src_content";
 
-    private static final String EVENT_IMPRESSION = "publisher.impression";
-    private static final String EVENT_CLICK = "publisher.click";
+    protected static final String EVENT_IMPRESSION = "publisher.impression";
+    protected static final String EVENT_CLICK = "publisher.click";
 
-    private static final String ON_SITE_TYPE = "onsitead_type";
-    private static final String ON_SITE_TYPE_PUBLISHER = "Publisher";
-    private static final String ON_SITE_ADVERTISER = "onsitead_advertiser";
-    private static final String ON_SITE_CAMPAIGN = "onsitead_campaign";
-    private static final String ON_SITE_FORMAT = "onsitead_format";
+    protected static final String ON_SITE_TYPE = "onsitead_type";
+    protected static final String ON_SITE_TYPE_PUBLISHER = "Publisher";
+    protected static final String ON_SITE_ADVERTISER = "onsitead_advertiser";
+    protected static final String ON_SITE_CAMPAIGN = "onsitead_campaign";
+    protected static final String ON_SITE_FORMAT = "onsitead_format";
 
     /**
      * Custom event name used when logging on Piano
-     *
+     * <p>
      *  Note: Must be add as custom event/property in your Piano Data Model
      */
     private static final String NOTIFICATION_DISPLAY_NAME = "batch_notification_display";
@@ -64,8 +61,8 @@ public class PianoDispatcher implements BatchEventDispatcher {
     private static final String MESSAGING_CLOSE_ERROR_NAME = "batch_in_app_close_error";
     private static final String MESSAGING_CLICK_NAME = "batch_in_app_click";
     private static final String MESSAGING_WEBVIEW_CLICK_NAME = "batch_in_app_webview_click";
-    private static final String BATCH_WEBVIEW_ANALYTICS_ID = "batch_webview_analytics_id";
-    private static final String BATCH_TRACKING_ID = "batch_tracking_id";
+    protected static final String BATCH_WEBVIEW_ANALYTICS_ID = "batch_webview_analytics_id";
+    protected static final String BATCH_TRACKING_ID = "batch_tracking_id";
     private static final String UNKNOWN_EVENT_NAME = "batch_unknown";
 
     /**
@@ -88,40 +85,26 @@ public class PianoDispatcher implements BatchEventDispatcher {
     private static final String UTM_CONTENT = "utm_content";
 
     /**
-     * Piano Analytics instance
-     */
-    private final PianoAnalytics pianoAnalytics;
-
-    /**
      * Whether Batch should send custom events (default: false)
-     *
+     * <p>
      * Note: Custom events must be defined in the Piano Data Model
      */
-    private boolean customEventsEnabled = false;
+    protected boolean customEventsEnabled = false;
 
     /**
      * Whether Batch should send onSiteAds events (default: true)
      */
-    private boolean onSiteAdsEventsEnabled = true;
+    protected boolean onSiteAdsEventsEnabled = true;
 
     /**
      * Whether Batch should handle UTM tags in campaign's deeplink
      * and custom payload. (default = true)
      */
-    private boolean isUTMTrackingEnabled = true;
-
-    /**
-     * Constructor
-     *
-     * @param context application context
-     */
-    public PianoDispatcher(@NonNull Context context) {
-        pianoAnalytics = PianoAnalytics.getInstance(context);
-    }
+    protected boolean isUTMTrackingEnabled = true;
 
     /**
      * Whether Batch should dispatch events as Piano Custom Event.
-     *
+     * <p>
      * Note: This method should be called as soon as possible in your Application subclass
      * just after your Piano Configuration.
      *
@@ -168,100 +151,10 @@ public class PianoDispatcher implements BatchEventDispatcher {
         return DISPATCHER_VERSION;
     }
 
-    /**
-     * Callback fired when a new Batch event is triggered
-     *
-     * @param type The type of the event
-     * @param payload The associated payload of the event
-     */
-    @Override
-    public void dispatchEvent(@NonNull Batch.EventDispatcher.Type type, @NonNull Batch.EventDispatcher.Payload payload) {
-
-        // Dispatch onSiteAds event
-        if (onSiteAdsEventsEnabled && shouldBeDispatchedAsOnSiteAd(type)) {
-            Event onSiteAdsEvent = buildPianoOnSiteAdsEvent(type, payload);
-            if (onSiteAdsEvent != null) {
-                pianoAnalytics.sendEvent(onSiteAdsEvent);
-            }
-        }
-
-        // Dispatch Custom Event if enabled
-        if (customEventsEnabled) {
-            Event event = buildPianoCustomEvent(type, payload);
-            pianoAnalytics.sendEvent(event);
-        }
-    }
-
-    /**
-     * Build an On-Site Ads Piano Event from a Batch Event
-     *
-     * @param type Batch event type
-     * @param payload Batch event payload
-     * @return The Piano event to send
-     */
-    @Nullable
-    @VisibleForTesting
-    Event buildPianoOnSiteAdsEvent(Batch.EventDispatcher.Type type, Batch.EventDispatcher.Payload payload) {
-
-        String pianoOnSiteEventName;
-        if (isImpression(type)) {
-            pianoOnSiteEventName = EVENT_IMPRESSION;
-        } else if (isClick(type)) {
-            pianoOnSiteEventName = EVENT_CLICK;
-        } else {
-            return null;
-        }
-
-        HashMap<String, Object> params = new HashMap<String, Object>() {{
-            put(ON_SITE_TYPE, ON_SITE_TYPE_PUBLISHER);
-            put(ON_SITE_ADVERTISER, getSource(payload));
-            put(ON_SITE_CAMPAIGN, getCampaign(payload));
-            put(ON_SITE_FORMAT, getMedium(payload, type));
-        }};
-       return new Event(pianoOnSiteEventName, params);
-    }
-
-    /**
-     * Build a Piano Custom Event from a Batch Event
-     *
-     * @param type Batch event type
-     * @param payload Batch event payload
-     * @return The Piano event to send
-     */
-    @VisibleForTesting
-    Event buildPianoCustomEvent(@NonNull Batch.EventDispatcher.Type type,
-                                  @NonNull Batch.EventDispatcher.Payload payload) {
-
-        String eventName = getPianoEventName(type);
-        HashMap<String, Object> eventData = new HashMap<String, Object>() {{
-            put(CAMPAIGN, getCampaign(payload));
-            put(MEDIUM, getMedium(payload, type));
-            put(SOURCE, getSource(payload));
-            put(SOURCE_FORCE, true);
-        }};
-
-        String trackingId = payload.getTrackingId();
-        if (trackingId != null && !trackingId.isEmpty()) {
-            eventData.put(BATCH_TRACKING_ID, trackingId);
-        }
-
-        String content = getContent(payload);
-        if (content != null && !content.isEmpty()) {
-            eventData.put(CONTENT, content);
-        }
-
-        if (type.isMessagingEvent()) {
-            String webViewAnalyticsId = payload.getWebViewAnalyticsID();
-            if (webViewAnalyticsId != null && !webViewAnalyticsId.isEmpty()) {
-                eventData.put(BATCH_WEBVIEW_ANALYTICS_ID, webViewAnalyticsId);
-            }
-        }
-        return new Event(eventName, eventData);
-    }
 
     /**
      * Get the campaign label to send at Piano.
-     *
+     * <p>
      * First, check for an "at_campaign" tag in the custom payload, or in the deeplink.
      * If not, check for an "utm_campaign" tag in the custom payload, or in the deeplink.
      * If not, check if there is a TrackingID attached.
@@ -290,7 +183,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
 
     /**
      * Get the medium
-     *
+     * <p>
      * Check for at_medium or utm_medium tags
      * If not found return "push" or "in-app" according to the batch event type
      *
@@ -299,7 +192,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
      * @return The medium
      */
     @NonNull
-    private String getMedium(@NonNull Batch.EventDispatcher.Payload payload, @NonNull Batch.EventDispatcher.Type type) {
+    protected String getMedium(@NonNull Batch.EventDispatcher.Payload payload, @NonNull Batch.EventDispatcher.Type type) {
         String medium;
         medium = getTagFromPayload(payload, AT_MEDIUM);
         if (medium != null && !medium.isEmpty()) {
@@ -319,7 +212,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
 
     /**
      * Get the source
-     *
+     * <p>
      * Check for utm_source tags
      * If not found return "Batch"
      *
@@ -327,7 +220,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
      * @return The source
      */
     @NonNull
-    private String getSource(@NonNull Batch.EventDispatcher.Payload payload) {
+    protected String getSource(@NonNull Batch.EventDispatcher.Payload payload) {
         String source = getTagFromPayload(payload, UTM_SOURCE);
         if (isUTMTrackingEnabled && source != null && !source.isEmpty()) {
             return source;
@@ -337,7 +230,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
 
     /**
      * Get the content
-     *
+     * <p>
      * Check for utm_content tag
      * If not found return null
      *
@@ -345,7 +238,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
      * @return The content
      */
     @Nullable
-    private String getContent(@NonNull Batch.EventDispatcher.Payload payload) {
+    protected String getContent(@NonNull Batch.EventDispatcher.Payload payload) {
         String content = getTagFromPayload(payload, UTM_CONTENT);
         if (isUTMTrackingEnabled && content != null && !content.isEmpty()) {
             return content;
@@ -360,7 +253,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
      * @return The corresponding event name for Piano
      */
     @NonNull
-    private String getPianoEventName(Batch.EventDispatcher.Type type) {
+    protected String getPianoEventName(Batch.EventDispatcher.Type type) {
         switch (type) {
             case NOTIFICATION_DISPLAY:
                 return NOTIFICATION_DISPLAY_NAME;
@@ -390,7 +283,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
      * @param type Batch event type
      * @return True if this kind of event should be dispatched as On-site Ads.
      */
-    private boolean shouldBeDispatchedAsOnSiteAd(Batch.EventDispatcher.Type type) {
+    protected boolean shouldBeDispatchedAsOnSiteAd(Batch.EventDispatcher.Type type) {
         return isImpression(type) || isClick(type);
     }
 
@@ -400,7 +293,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
      * @param type Batch event type
      * @return True if it's an impression
      */
-    private boolean isImpression(Batch.EventDispatcher.Type type) {
+    protected boolean isImpression(Batch.EventDispatcher.Type type) {
         return type.equals(Batch.EventDispatcher.Type.NOTIFICATION_DISPLAY) ||
                 type.equals(Batch.EventDispatcher.Type.MESSAGING_SHOW);
     }
@@ -411,7 +304,7 @@ public class PianoDispatcher implements BatchEventDispatcher {
      * @param type Batch event type
      * @return True if it's a click
      */
-    private boolean isClick(Batch.EventDispatcher.Type type) {
+    protected boolean isClick(Batch.EventDispatcher.Type type) {
         return type.equals(Batch.EventDispatcher.Type.NOTIFICATION_OPEN) ||
                 type.equals(Batch.EventDispatcher.Type.MESSAGING_CLICK) ||
                 type.equals(Batch.EventDispatcher.Type.MESSAGING_WEBVIEW_CLICK);
@@ -489,3 +382,4 @@ public class PianoDispatcher implements BatchEventDispatcher {
         return map;
     }
 }
+
